requirements.txt
crewai
langchain
langchain-community
langchain-openai
openai
python-decouple
duckduckgo-search
markdown